+++ .gitignore
@@ -2,5 +2,6 @@
 *.ipr
 *.iml
 *.iws
-/build/
+build/
 /.gradle/
+out/
