+++ settings.gradle
@@ -1,2 +1,3 @@
-rootProject.name = 'spinnaker-gradle-project'
+rootProject.name = 'spinnaker-gradle'
 
+include 'spinnaker-dev-plugin', 'spinnaker-publishing-plugin', 'spinnaker-gradle-project'
