+++ spinnaker-dev-plugin/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+  compile 'org.yaml:snakeyaml:1.21'
+}
+
