+++ spinnaker-dev-plugin/src/main/groovy/com/netflix/spinnaker/gradle/baseproject/SpinnakerBaseProjectPlugin.groovy
@@ -1,0 +1,15 @@
+package com.netflix.spinnaker.gradle.baseproject
+
+import com.netflix.spinnaker.gradle.dependency.SpinnakerDependencyPlugin
+import com.netflix.spinnaker.gradle.idea.SpinnakerIdeaConfigPlugin
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+
+class SpinnakerBaseProjectPlugin implements Plugin<Project> {
+    @Override
+    void apply(Project project) {
+        project.plugins.apply(SpinnakerBaseProjectConventionsPlugin)
+        project.plugins.apply(SpinnakerIdeaConfigPlugin)
+        project.plugins.apply(SpinnakerDependencyPlugin)
+    }
+}
