+++ spinnaker-dev-plugin/src/main/resources/META-INF/gradle-plugins/spinnaker.dependency.properties
