+++ spinnaker-publishing-plugin/build.gradle
@@ -1,0 +1,6 @@
+dependencies {
+  compile project(":spinnaker-dev-plugin")
+  compile 'com.netflix.nebula:gradle-netflixoss-project-plugin:5.0.0'
+  compile 'com.netflix.nebula:gradle-ospackage-plugin:4.8.0'
+  compile 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.0'
+}
