+++ spinnaker-publishing-plugin/src/main/groovy/com/netflix/spinnaker/gradle/ospackage/OspackageBintrayPublishPlugin.groovy
